+++ docs/_pages/releases.md
@@ -8,6 +8,10 @@ sidebar:
 ---
 ## Unreleased
 
+### What's new
+
+* Added overload for collection's `Contain` to assert the occurrences - [#238](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/238)
+
 ### Improvements
 * Always use improved string difference visualization independent of the subject's or expectation's string length - [#215](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/215)